+++ src/main/java/core/mvc/tobe/RequestHandlerAdapter.java
@@ -2,8 +2,8 @@ package core.mvc.tobe;
 
 import core.mvc.ModelAndView;
 import core.mvc.asis.Controller;
-import core.mvc.tobe.resolver.StringPathViewResolver;
-import core.mvc.tobe.resolver.ViewResolver;
+import core.mvc.tobe.resolver.view.StringPathViewResolver;
+import core.mvc.tobe.resolver.view.ViewResolver;
 import core.mvc.tobe.view.View;
 
 import javax.servlet.http.HttpServletRequest;
