+++ src/test/java/core/mvc/tobe/HandlerMethodArgumentResolverTest.java
@@ -3,8 +3,8 @@ package core.mvc.tobe;
 import core.annotation.web.PathVariable;
 import core.annotation.web.RequestMapping;
 import core.mvc.ModelAndView;
-import core.mvc.tobe.resolver.ArgumentResolvers;
-import core.mvc.tobe.resolver.HandlerMethodArgumentResolver;
+import core.mvc.tobe.resolver.argument.ArgumentResolvers;
+import core.mvc.tobe.resolver.argument.HandlerMethodArgumentResolver;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
